Aula 299: Lidando com retornos

+++ app/Http/Controllers/MarcaController.php
@@ -57,6 +57,12 @@ class MarcaController extends Controller
      */
     public function show($id)
     {   
+        $marca = $this->marca->find($id);
+
+        if($marca === null){
+            return ['msg'=>'Dados não encontrados'];
+        }
+
         return $this->marca->find($id);
     }
 
@@ -81,6 +87,9 @@ class MarcaController extends Controller
     public function update(Request $request, $id)
     {
         $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['msg'=>'Dados não encontrados, não é possível atualizar'];
+        }
         $marca->update($request->all());
         return $marca;
     }
@@ -94,6 +103,9 @@ class MarcaController extends Controller
     public function destroy($id)
     {   
         $marca = $this->marca->find($id);
+        if($marca === null){
+            return ['msg'=>'Dados não encontrados, não é possível excluir'];
+        }
         $marca->delete($id);
         return ['msg'=>'A marca foi removida com sucesso!'];
     }
